+++ sqlbuilder_insert.go
@@ -61,22 +61,28 @@ func (ib *InsertBuilder) SetMap(m map[string]any, opts ...BuilderOption) *Insert
 		return ib
 	}
 
-	bo := &BuilderOptions{
-		DbColumns: make(map[string]bool),
-	}
+	bo := &BuilderOptions{}
 	for _, opt := range opts {
 		opt(bo)
 	}
 
 	for n, v := range m {
 		if bo.ToSnake != nil {
-			n = bo.ToSnake(n)
+			sn := bo.ToSnake(n)
+			if sn != n {
+				delete(m, n)
+				m[sn] = v
+			}
 		}
+	}
 
-		if _, ok := bo.DbColumns[n]; ok {
+	for _, n := range bo.Columns {
+		v, ok := m[n]
+		if ok {
 			ib.Set(n, v)
 		}
 	}
+
 	return ib
 }